removed ExpectationKwargs and updated tests
