+++ tests/core/test_expectation_suite.py
@@ -3,7 +3,6 @@ from copy import copy, deepcopy
 import pytest
 
 from great_expectations.core import ExpectationConfiguration, ExpectationSuite
-from great_expectations.exceptions import DataContextError
 
 
 @pytest.fixture
@@ -38,42 +37,6 @@ def exp4():
     return ExpectationConfiguration(
         expectation_type="expect_column_values_to_be_in_set",
         kwargs={"column": "b", "value_set": [1, 2, 3], "result_format": "BASIC"},
-        meta={"notes": "This is an expectation."},
-    )
-
-
-@pytest.fixture
-def exp5():
-    return ExpectationConfiguration(
-        expectation_type="expect_column_values_to_be_in_set",
-        kwargs={"column": "b", "value_set": [1, 2, 3], "result_format": "COMPLETE"},
-        meta={"notes": "This is an expectation."},
-    )
-
-
-@pytest.fixture
-def exp6():
-    return ExpectationConfiguration(
-        expectation_type="expect_column_values_to_be_in_set",
-        kwargs={"column": "b", "value_set": [1, 2]},
-        meta={"notes": "This is an expectation."},
-    )
-
-
-@pytest.fixture
-def exp7():
-    return ExpectationConfiguration(
-        expectation_type="expect_column_values_to_be_in_set",
-        kwargs={"column": "b", "value_set": [1, 2, 3, 4]},
-        meta={"notes": "This is an expectation."},
-    )
-
-
-@pytest.fixture
-def exp8():
-    return ExpectationConfiguration(
-        expectation_type="expect_column_values_to_be_in_set",
-        kwargs={"column": "b", "value_set": [1, 2, 3]},
         meta={"notes": "This is an expectation."},
     )
 
@@ -124,6 +87,15 @@ def empty_suite():
 
 
 @pytest.fixture
+def single_expectation_suite(exp1):
+    return ExpectationSuite(
+        expectation_suite_name="warning",
+        expectations=[exp1],
+        meta={"notes": "This is an expectation suite."},
+    )
+
+
+@pytest.fixture
 def suite_with_table_and_column_expectations(
     exp1, exp2, exp3, exp4, column_pair_expectation, table_exp1, table_exp2, table_exp3
 ):
@@ -153,43 +125,12 @@ def suite_with_table_and_column_expectations(
     ]
     return suite
 
-@pytest.fixture
-def suite_with_column_pair_and_table_expectations(
-    table_exp1, table_exp2, table_exp3, column_pair_expectation
-):
-    suite = ExpectationSuite(
-        expectation_suite_name="warning",
-        expectations=[
-            column_pair_expectation,
-            table_exp1,
-            table_exp2,
-            table_exp3,
-        ],
-        meta={"notes": "This is an expectation suite."},
-    )
-    assert suite.expectations == [
-        column_pair_expectation,
-        table_exp1,
-        table_exp2,
-        table_exp3,
-    ]
-    return suite
-
 
 @pytest.fixture
 def baseline_suite(exp1, exp2):
     return ExpectationSuite(
         expectation_suite_name="warning",
         expectations=[exp1, exp2],
-        meta={"notes": "This is an expectation suite."},
-    )
-
-
-@pytest.fixture
-def domain_success_runtime_suite(exp1, exp2, exp3, exp4, exp5):
-    return ExpectationSuite(
-        expectation_suite_name="warning",
-        expectations=[exp1, exp2, exp3, exp4, exp5],
         meta={"notes": "This is an expectation suite."},
     )
 
@@ -219,15 +160,6 @@ def different_suite(exp1, exp4):
     return ExpectationSuite(
         expectation_suite_name="warning",
         expectations=[exp1, exp4],
-        meta={"notes": "This is an expectation suite."},
-    )
-
-
-@pytest.fixture
-def single_expectation_suite(exp1):
-    return ExpectationSuite(
-        expectation_suite_name="warning",
-        expectations=[exp1],
         meta={"notes": "This is an expectation suite."},
     )
 
@@ -493,124 +425,3 @@ def test_get_column_expectations(
     assert obs == [exp1, exp2, exp3, exp4]
 
 
-def test_find_expectation_indexes(
-    exp1, exp2, exp3, exp4, exp5, domain_success_runtime_suite
-):
-    assert domain_success_runtime_suite.find_expectation_indexes(exp4, "domain") == [1, 2, 3, 4]
-    assert domain_success_runtime_suite.find_expectation_indexes(exp4, "success") == [3, 4]
-    assert domain_success_runtime_suite.find_expectation_indexes(exp4, "runtime") == [3]
-
-
-def test_remove_expectations(
-    exp1, exp2, exp3, exp4, exp5, single_expectation_suite, domain_success_runtime_suite
-):
-    domain_success_runtime_suite.remove_expectation(exp5, match_type="runtime", remove_multiple_matches=False) # remove one matching expectation
-
-    with pytest.raises(ValueError):
-        domain_success_runtime_suite.remove_expectation(exp5, match_type="runtime")
-    assert domain_success_runtime_suite.find_expectation_indexes(exp4, "domain") == [
-        1,
-        2,
-        3,
-    ]
-    assert domain_success_runtime_suite.find_expectation_indexes(exp4, "success") == [3]
-
-    with pytest.raises(ValueError):
-        domain_success_runtime_suite.remove_expectation(exp4, match_type="domain", remove_multiple_matches=False)
-
-    # remove 3 matching expectations
-    domain_success_runtime_suite.remove_expectation(exp4, match_type="domain", remove_multiple_matches=True)
-
-    with pytest.raises(ValueError):
-        domain_success_runtime_suite.remove_expectation(exp2, match_type="runtime")
-    with pytest.raises(ValueError):
-        domain_success_runtime_suite.remove_expectation(exp3, match_type="runtime")
-
-    assert domain_success_runtime_suite.find_expectation_indexes(
-        exp1, match_type="domain"
-    ) == [0]
-    assert domain_success_runtime_suite.isEquivalentTo(single_expectation_suite)
-
-
-def test_patch_expectation_replace(exp5, exp6, domain_success_runtime_suite):
-    assert domain_success_runtime_suite.expectations[4] is exp5
-
-    assert not domain_success_runtime_suite.expectations[4].isEquivalentTo(
-        exp6, match_type="success"
-    )
-    domain_success_runtime_suite.patch(
-        exp5,
-        op="replace",
-        path="/value_set",
-        value=[1, 2],
-        match_type="runtime",
-    )
-    assert domain_success_runtime_suite.expectations[4].isEquivalentTo(
-        exp6, match_type="success"
-    )
-
-
-def test_patch_expectation_add(exp5, exp7, domain_success_runtime_suite):
-    assert domain_success_runtime_suite.expectations[4] is exp5
-
-    assert not domain_success_runtime_suite.expectations[4].isEquivalentTo(
-        exp7, match_type="success"
-    )
-    domain_success_runtime_suite.patch(
-        exp5,
-        op="add",
-        path="/value_set/-",
-        value=4,
-        match_type="runtime",
-    )
-    assert domain_success_runtime_suite.expectations[4].isEquivalentTo(
-        exp7, match_type="success"
-    )
-
-
-def test_patch_expectation_remove(exp5, exp8, domain_success_runtime_suite):
-    assert domain_success_runtime_suite.expectations[4] is exp5
-
-    assert not domain_success_runtime_suite.expectations[4].isEquivalentTo(
-        exp8, match_type="runtime"
-    )
-    domain_success_runtime_suite.patch(
-        exp5,
-        op="remove",
-        path="/result_format",
-        value=None,
-        match_type="runtime",
-    )
-    assert domain_success_runtime_suite.expectations[4].isEquivalentTo(
-        exp8, match_type="runtime"
-    )
-
-
-def test_add_expectation(
-    exp1, exp2, exp4, single_expectation_suite, baseline_suite, different_suite, domain_success_runtime_suite
-):
-    assert not single_expectation_suite.isEquivalentTo(baseline_suite)
-    single_expectation_suite.add_expectation(exp2, match_type="runtime", overwrite_existing=False)
-    assert single_expectation_suite.isEquivalentTo(baseline_suite)
-
-    # Should raise if overwrite_existing=False and a matching expectation is found
-    with pytest.raises(DataContextError):
-        single_expectation_suite.add_expectation(exp4, match_type="domain", overwrite_existing=False)
-
-    assert not single_expectation_suite.isEquivalentTo(different_suite)
-    single_expectation_suite.add_expectation(exp4, match_type="domain", overwrite_existing=True)
-    assert single_expectation_suite.isEquivalentTo(different_suite)
-
-    # Should raise if more than one matching expectation is found
-    with pytest.raises(ValueError):
-        domain_success_runtime_suite.add_expectation(exp2, match_type="success", overwrite_existing=False)
-
-
-def test_remove_all_expectations_of_type(
-    suite_with_table_and_column_expectations, suite_with_column_pair_and_table_expectations
-):
-    assert not suite_with_table_and_column_expectations.isEquivalentTo(suite_with_column_pair_and_table_expectations)
-
-    suite_with_table_and_column_expectations.remove_all_expectations_of_type("expect_column_values_to_be_in_set")
-
-    assert suite_with_table_and_column_expectations.isEquivalentTo(suite_with_column_pair_and_table_expectations)
