+++ tests/data_asset/test_data_asset.py
@@ -3,7 +3,7 @@ import pytest
 from great_expectations import __version__ as ge_version
 from great_expectations.core import (
     ExpectationConfiguration,
-    ExpectationKwargs,
+    # ExpectationKwargs,
     ExpectationSuite,
 )
 from great_expectations.data_asset import DataAsset, FileDataAsset
